Now presenting nice suggestions!
After filtering for various things it is now possible to produce a nice
list of options.

Here are the 10 least used presented according to my keyboard layout
together with how often they occur in english words.

    Ctrl + K 12
    Ctrl + R 11
    Shift + X 10
    Ctrl + D 9
    Ctrl + J 8
    Ctrl + C 5
    Ctrl + S 3
    Ctrl + W 2
    Alt + X 1
    Ctrl + V 1

Alt + X and Ctrl + V only occur once. These correspond to kx and jv.
While there are 2 Alts and Ctrls there is a filtering rule saying it
must use both sides of the keyboard. hence Alt + X is not dx for
instance.

Next step would be to filter out combinations, like Ctrl + V, that are
already in use.